requirements.txt
pandas
yeelights
